RiseBackend/requirements.txt
Django==5.1.6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
